+++ addons/ammunition/niarms/CfgMagazineWells.hpp
@@ -1,0 +1,71 @@
+class CfgMagazineWells {
+    // Handguns
+    class CBA_9x19_P226 {
+        ADDON[] = {
+            QCLASS(15Rnd_9x19_P226_Ball)
+        };
+    };
+
+    class CBA_9x19_P228 {
+        ADDON[] = {
+            QCLASS(15Rnd_9x19_P226_Ball)
+        };
+    };
+
+    class CBA_9x19_P239 {
+        ADDON[] = {
+            QCLASS(10Rnd_9x19_P239_Ball)
+        };
+    };
+    class CBA_357SIG_P239 {
+        ADDON[] = {
+            QCLASS(8Rnd_357_P239_Magnum)
+        };
+    };
+
+    class CBA_357SIG_P226 {
+        ADDON[] = {
+            QCLASS(12Rnd_357_P226_Magnum)
+        };
+    };
+
+    // Rifles
+    class CBA_300BLK_STANAG {
+        ADDON[] = {
+            QCLASS(30Rnd_300AAC_Ball),
+            QCLASS(30Rnd_300AAC_Ball_Tracer),
+            QCLASS(30Rnd_300AAC_EPR),
+            QCLASS(30Rnd_300AAC_AP)
+        };
+    };
+
+    class CBA_762x51_LINKS {
+        ADDON[] = {
+            QCLASS(100Rnd_762x51_Ball_Belt),
+            QCLASS(100Rnd_762x51_Ball_Tracer_Belt),
+            QCLASS(100Rnd_762x51_EPR_Belt)
+        };
+    };
+
+    class CBA_556x45_SG550 {
+        ADDON[] = {
+            QCLASS(30Rnd_556x45_Ball_SG),
+            QCLASS(30Rnd_556x45_Ball_Tracer_SG),
+            QCLASS(30Rnd_556x45_EPR_SG),
+            QCLASS(30Rnd_556x45_AP_SG)
+        };
+    };
+
+    class CBA_556x45_STEYR {
+        ADDON[] = {
+            QCLASS(30Rnd_556x45_Ball_AUG),
+            QCLASS(30Rnd_556x45_Ball_Tracer_AUG),
+            QCLASS(30Rnd_556x45_EPR_AUG),
+            QCLASS(30Rnd_556x45_AP_AUG),
+            QCLASS(42Rnd_556x45_Ball_AUG),
+            QCLASS(42Rnd_556x45_Ball_Tracer_AUG),
+            QCLASS(42Rnd_556x45_EPR_AUG),
+            QCLASS(42Rnd_556x45_AP_AUG)
+        };
+    };
+};
