--- addons/ammunition/ws/CfgWeapons.hpp
@@ -1,12 +1,0 @@
-class CfgWeapons {
-    // Western Sahara Weapons
-    class arifle_Galat_base_lxWS;
-    class arifle_Galat_lxWS: arifle_Galat_base_lxWS {
-        magazineWell[] += {"CBA_762x39_AK"};
-    };
-
-    class arifle_SPAR_01_base_F;
-    class arifle_XMS_Base_lxWS: arifle_SPAR_01_base_F {
-        magazineWell[] += {"CBA_556x45_STANAG"};
-    };
-};
