+++ addons/ammunition/ws/config.cpp
@@ -20,4 +20,3 @@ class CfgPatches {
 };
 
 #include "CfgMagazines.hpp"
-#include "CfgWeapons.hpp"
